requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-docx>=0.8.11
openai>=1.0.0
requests>=2.31.0
